refactor: split large source files
