+++ Sources/DefiDaemon/DaemonFocus.swift
@@ -420,503 +420,4 @@ extension Daemon {
     }
   }
 
-  func requeueDisplacedPointerFocusAfterDisplayChange(
-    _ recovery: DisplacedPointerFocusRecovery
-  ) {
-    switch recovery {
-    case .command(let request, let timestamp):
-      guard let monitorID = state.reboundFocusMonitorID(
-        for: request.windowID
-      ) else { return }
-      pendingAnimatedFocus = PendingAnimatedFocus(
-        windowID: request.windowID,
-        previousSelectedWindowID: request.previousSelectedWindowID,
-        monitorID: monitorID,
-        sourceWorkspaceID: request.sourceWorkspaceID,
-        commandGeneration: request.commandGeneration,
-        focusInputTimestamp: pointerDisplacedFocusInputTimestamp(
-          commandInputTimestamp: request.focusInputTimestamp,
-          pointerInputTimestamp: timestamp
-        ),
-        cursorWarpInputTimestamp: nil,
-        retryCount: request.retryCount
-      )
-    case .workspace(let request, let timestamp):
-      guard let monitorID = state.reboundFocusMonitorID(
-        for: request.requestedWindowID,
-        requestedWorkspaceID: request.requestedWorkspaceID
-      ) else { return }
-      pendingWorkspaceFocus = PendingWorkspaceFocus(
-        monitorID: monitorID,
-        requestedWorkspaceID: request.requestedWorkspaceID,
-        previousWorkspaceID: request.previousWorkspaceID,
-        requestedWindowID: request.requestedWindowID,
-        restoresPreviousWorkspaceOnCancellation:
-          request.restoresPreviousWorkspaceOnCancellation,
-        commandGeneration: request.commandGeneration,
-        focusInputTimestamp: pointerDisplacedFocusInputTimestamp(
-          commandInputTimestamp: request.focusInputTimestamp,
-          pointerInputTimestamp: timestamp
-        ),
-        cursorWarpInputTimestamp: nil,
-        retryCount: request.retryCount
-      )
-      submittedWorkspaceFocusGeneration = nil
-    }
-  }
-
-  func requeuePreservedFocusAfterMonitorRetention(
-    command: PendingAnimatedFocus?,
-    workspace: PendingWorkspaceFocus?,
-    displaced: DisplacedPointerFocusRecovery?
-  ) {
-    if let displaced {
-      requeueDisplacedPointerFocusAfterDisplayChange(displaced)
-      return
-    }
-    if let command,
-      let monitorID = state.reboundFocusMonitorID(for: command.windowID)
-    {
-      pendingAnimatedFocus = PendingAnimatedFocus(
-        windowID: command.windowID,
-        previousSelectedWindowID: command.previousSelectedWindowID,
-        monitorID: monitorID,
-        sourceWorkspaceID: command.sourceWorkspaceID,
-        commandGeneration: command.commandGeneration,
-        focusInputTimestamp: command.focusInputTimestamp,
-        cursorWarpInputTimestamp: command.cursorWarpInputTimestamp,
-        retryCount: command.retryCount
-      )
-    }
-    if let workspace,
-      let monitorID = state.reboundFocusMonitorID(
-        for: workspace.requestedWindowID,
-        requestedWorkspaceID: workspace.requestedWorkspaceID
-      )
-    {
-      pendingWorkspaceFocus = PendingWorkspaceFocus(
-        monitorID: monitorID,
-        requestedWorkspaceID: workspace.requestedWorkspaceID,
-        previousWorkspaceID: workspace.previousWorkspaceID,
-        requestedWindowID: workspace.requestedWindowID,
-        restoresPreviousWorkspaceOnCancellation:
-          workspace.restoresPreviousWorkspaceOnCancellation,
-        commandGeneration: workspace.commandGeneration,
-        focusInputTimestamp: workspace.focusInputTimestamp,
-        cursorWarpInputTimestamp: workspace.cursorWarpInputTimestamp,
-        retryCount: workspace.retryCount
-      )
-      submittedWorkspaceFocusGeneration = nil
-    }
-  }
-
-  func rearmPointerFocusTransition() {
-    lastPointerWindowID = nil
-    lastRawPointerWindowID = nil
-    hotKeys?.resetPointerWindowTransition()
-  }
-
-  func invalidateSubmittedCommandFocus(
-    recoveringTo windowID: WindowID? = nil
-  ) {
-    guard let requestID = submittedCommandFocusRequestID else {
-      submittedCommandFocusRequestTimestamp = nil
-      submittedCommandFocusRecoveryGeneration = nil
-      submittedCommandFocus = nil
-      return
-    }
-    let timestamp = submittedCommandFocus?.focusInputTimestamp
-      ?? submittedCommandFocusRequestTimestamp
-    let cancelled = platform.cancelFocus(requestID, recoveringTo: windowID)
-    if !cancelled, let windowID, let timestamp {
-      nextCommandFocusRecoveryGeneration &+= 1
-      let recoveryGeneration = nextCommandFocusRecoveryGeneration
-      let recoveryID = platform.focus(
-        windowID,
-        unlessUserInputAfter: timestamp,
-        completion: { [weak self] _ in
-          guard let self,
-            self.submittedCommandFocusRecoveryGeneration
-              == recoveryGeneration
-          else { return }
-          self.submittedCommandFocusRequestID = nil
-          self.submittedCommandFocusRequestTimestamp = nil
-          self.submittedCommandFocusRecoveryGeneration = nil
-        }
-      )
-      submittedCommandFocusRequestID = recoveryID
-      submittedCommandFocusRequestTimestamp =
-        recoveryID == nil ? nil : timestamp
-      submittedCommandFocusRecoveryGeneration =
-        recoveryID == nil ? nil : recoveryGeneration
-    } else {
-      submittedCommandFocusRequestID = nil
-      submittedCommandFocusRequestTimestamp = nil
-      submittedCommandFocusRecoveryGeneration = nil
-    }
-    submittedCommandFocus = nil
-  }
-
-  func invalidateSubmittedWorkspaceFocus(
-    recoveringTo windowID: WindowID? = nil
-  ) {
-    guard let requestID = submittedWorkspaceFocusRequestID else {
-      submittedWorkspaceFocusRequestTimestamp = nil
-      submittedWorkspaceFocusGeneration = nil
-      submittedWorkspaceFocusRecoveryGeneration = nil
-      return
-    }
-    let timestamp = pendingWorkspaceFocus?.focusInputTimestamp
-      ?? submittedWorkspaceFocusRequestTimestamp
-    let cancelled = platform.cancelFocus(requestID, recoveringTo: windowID)
-    if !cancelled, let windowID, let timestamp {
-      nextWorkspaceFocusRecoveryGeneration &+= 1
-      let recoveryGeneration = nextWorkspaceFocusRecoveryGeneration
-      let recoveryID = platform.focus(
-        windowID,
-        unlessUserInputAfter: timestamp,
-        completion: { [weak self] _ in
-          guard let self,
-            self.submittedWorkspaceFocusRecoveryGeneration
-              == recoveryGeneration
-          else { return }
-          self.submittedWorkspaceFocusRequestID = nil
-          self.submittedWorkspaceFocusRequestTimestamp = nil
-          self.submittedWorkspaceFocusRecoveryGeneration = nil
-        }
-      )
-      submittedWorkspaceFocusRequestID = recoveryID
-      submittedWorkspaceFocusRequestTimestamp =
-        recoveryID == nil ? nil : timestamp
-      submittedWorkspaceFocusRecoveryGeneration =
-        recoveryID == nil ? nil : recoveryGeneration
-    } else {
-      submittedWorkspaceFocusRequestID = nil
-      submittedWorkspaceFocusRequestTimestamp = nil
-      submittedWorkspaceFocusRecoveryGeneration = nil
-    }
-    submittedWorkspaceFocusGeneration = nil
-  }
-
-  func invalidatePointerFocusIntent(recoveringTo windowID: WindowID? = nil) {
-    pointerFocusGeneration &+= 1
-    pendingPointerFocus = nil
-    invalidateSubmittedPointerFocusRecovery(recoveringTo: windowID)
-    if let submittedPointerFocusRequestID {
-      let timestamp = submittedPointerFocusTimestamp
-      let cancelled = platform.cancelFocus(
-        submittedPointerFocusRequestID,
-        recoveringTo: windowID
-      )
-      if let recoveryWindowID = pointerFocusRecoveryTargetAfterCancellation(
-        cancellationSucceeded: cancelled,
-        logicalFocusWindowID: windowID
-      ), let timestamp
-      {
-        recoverPointerFocus(
-          to: recoveryWindowID,
-          unlessUserInputAfter: timestamp
-        )
-      }
-      self.submittedPointerFocusRequestID = nil
-      self.submittedPointerFocusTimestamp = nil
-      self.submittedPointerFocusGeneration = nil
-    }
-  }
-
-  private func recoverPointerFocus(
-    to windowID: WindowID?,
-    unlessUserInputAfter timestamp: TimeInterval
-  ) {
-    guard let windowID else { return }
-    nextPointerFocusRecoveryGeneration &+= 1
-    let recoveryGeneration = nextPointerFocusRecoveryGeneration
-    let recoveryID = platform.focus(
-      windowID,
-      unlessUserInputAfter: timestamp,
-      completion: { [weak self] _ in
-        guard let self,
-          self.submittedPointerFocusRecoveryGeneration == recoveryGeneration
-        else { return }
-        self.submittedPointerFocusRecoveryRequestID = nil
-        self.submittedPointerFocusRecoveryTimestamp = nil
-        self.submittedPointerFocusRecoveryGeneration = nil
-      }
-    )
-    submittedPointerFocusRecoveryRequestID = recoveryID
-    submittedPointerFocusRecoveryTimestamp =
-      recoveryID == nil ? nil : timestamp
-    submittedPointerFocusRecoveryGeneration =
-      recoveryID == nil ? nil : recoveryGeneration
-  }
-
-  private func invalidateSubmittedPointerFocusRecovery(
-    recoveringTo windowID: WindowID?
-  ) {
-    guard let recoveryRequestID = submittedPointerFocusRecoveryRequestID else {
-      submittedPointerFocusRecoveryTimestamp = nil
-      submittedPointerFocusRecoveryGeneration = nil
-      return
-    }
-    let timestamp = submittedPointerFocusRecoveryTimestamp
-    let recoveryGeneration = submittedPointerFocusRecoveryGeneration
-    let cancelled = platform.cancelFocus(
-      recoveryRequestID,
-      recoveringTo: windowID
-    )
-    if !cancelled, let windowID, let timestamp {
-      recoverPointerFocus(
-        to: windowID,
-        unlessUserInputAfter: timestamp
-      )
-    }
-    guard submittedPointerFocusRecoveryGeneration == recoveryGeneration else {
-      return
-    }
-    submittedPointerFocusRecoveryRequestID = nil
-    submittedPointerFocusRecoveryTimestamp = nil
-    submittedPointerFocusRecoveryGeneration = nil
-  }
-
-  private func pointerFocusIsReady(for windowID: WindowID) -> Bool {
-    guard let targetMonitorID = state.monitorID(containing: windowID) else {
-      return false
-    }
-    return focusMonitorIsReady(
-      targetMonitorID: targetMonitorID,
-      scrollingMonitorIDs: Set(scrollAnimations.keys.map(\.monitorID)),
-      pendingFrameMonitorIDs: Set(
-        platform.pendingFrameWindowIDs.compactMap {
-          state.monitorID(containing: $0)
-        }
-      ),
-      deferredSlowMonitorIDs: Set(
-        deferredSlowWindowIDs.compactMap {
-          state.monitorID(containing: $0)
-        }
-      )
-    )
-  }
-
-  private func cancellationKeepsRequestedWindow(
-    _ windowID: WindowID,
-    requestInputTimestamp: TimeInterval
-  ) -> Bool {
-    return cancelledFocusTargetsRequestedWindow(
-      requestedWindowID: windowID,
-      requestedWindowIsNativelyFocused:
-        platform.isWindowNativelyFocused(windowID),
-      cancellingFocusTargetWindowID: platform.userInputTracker
-        .focusRecoveryTarget(after: requestInputTimestamp)?.windowID
-    )
-  }
-
-  private func focusCompletionRequiresLogicalRollback(
-    _ result: NativeFocusResult
-  ) -> Bool {
-    switch result {
-    case .failedAfterMutation, .cancelledAfterInputMutation:
-      true
-    case .completed, .completedWithoutMutation, .frameSuperseded,
-      .superseded, .supersededAfterMutation, .cancelled,
-      .cancelledAfterMutation, .failed:
-      false
-    }
-  }
-
-  func commitCommandFocus(
-    _ windowID: WindowID,
-    previousSelectedWindowID: WindowID?,
-    monitorID: MonitorID,
-    sourceWorkspaceID: WorkspaceID,
-    commandGeneration: UInt64,
-    focusInputTimestamp: TimeInterval,
-    cursorWarpInputTimestamp: TimeInterval?,
-    retryCount: Int = 0
-  ) {
-    invalidateSubmittedCommandFocus()
-    let request = PendingAnimatedFocus(
-      windowID: windowID,
-      previousSelectedWindowID: previousSelectedWindowID,
-      monitorID: monitorID,
-      sourceWorkspaceID: sourceWorkspaceID,
-      commandGeneration: commandGeneration,
-      focusInputTimestamp: focusInputTimestamp,
-      cursorWarpInputTimestamp: cursorWarpInputTimestamp,
-      retryCount: retryCount
-    )
-    submittedCommandFocus = request
-    let committedCursorWarpInputTimestamp =
-      platform.cursorWarpFrameIsReady(for: windowID)
-      ? cursorWarpInputTimestamp
-      : nil
-    submittedCommandFocusRequestID = platform.focus(
-      windowID,
-      unlessUserInputAfter: focusInputTimestamp,
-      cursorWarpUnlessPointerMovedAfter: committedCursorWarpInputTimestamp,
-      cursorWarpIsCurrent: { [weak self] in
-        guard let self else { return false }
-        return self.submittedCommandFocus?.windowID == request.windowID
-          && self.submittedCommandFocus?.commandGeneration
-            == request.commandGeneration
-      },
-      completion: { [weak self] result in
-        guard let self else { return }
-        guard commandFocusCompletionIsCurrent(
-          submittedWindowID: self.submittedCommandFocus?.windowID,
-          submittedGeneration:
-            self.submittedCommandFocus?.commandGeneration,
-          completedWindowID: request.windowID,
-          completedGeneration: request.commandGeneration
-        ) else { return }
-        self.submittedCommandFocus = nil
-        self.submittedCommandFocusRequestID = nil
-        self.submittedCommandFocusRequestTimestamp = nil
-        self.submittedCommandFocusRecoveryGeneration = nil
-        if result == .failed || result == .failedAfterMutation,
-          let nextRetryCount = nextCommandFocusRetryCount(
-            currentRetryCount: request.retryCount,
-            maximumRetryCount: 1,
-            requestGeneration: request.commandGeneration,
-            currentGeneration: self.commandGeneration,
-            requestedWindowID: request.windowID,
-            selectedWindowID: self.state.selectedWindowID(on: request.monitorID)
-          )
-        {
-          self.pendingAnimatedFocus = PendingAnimatedFocus(
-            windowID: request.windowID,
-            previousSelectedWindowID: request.previousSelectedWindowID,
-            monitorID: request.monitorID,
-            sourceWorkspaceID: request.sourceWorkspaceID,
-            commandGeneration: request.commandGeneration,
-            focusInputTimestamp: request.focusInputTimestamp,
-            cursorWarpInputTimestamp: request.cursorWarpInputTimestamp,
-            retryCount: nextRetryCount
-          )
-          return
-        }
-        guard !self.cancellationKeepsRequestedWindow(
-          windowID,
-          requestInputTimestamp: focusInputTimestamp
-        ),
-          let fallbackWindowID = commandFocusCancellationFallback(
-            cancelledBeforeMutation:
-              result == .cancelled || result == .failed,
-            rollbackAfterMutation:
-              self.focusCompletionRequiresLogicalRollback(result),
-            requestGeneration: commandGeneration,
-            currentGeneration: self.commandGeneration,
-            requestedWindowID: windowID,
-            selectedWindowID: self.state.selectedWindowID(on: monitorID),
-            previousSelectedWindowID: previousSelectedWindowID,
-            sourceWorkspaceID: sourceWorkspaceID,
-            previousSelectedWindowWorkspaceID:
-              previousSelectedWindowID.flatMap {
-                self.state.location(containing: $0)?.workspaceID
-              }
-          ),
-          self.state.location(containing: fallbackWindowID)?.monitorID == monitorID
-        else {
-          return
-        }
-        _ = focusWindow(fallbackWindowID, state: &self.state)
-        self.activeMonitorID = self.state.monitorID(
-          containing: fallbackWindowID
-        )
-        self.needsDesktopSync = true
-        self.updateMenuBar()
-      }
-    )
-    submittedCommandFocusRequestTimestamp =
-      submittedCommandFocusRequestID == nil ? nil : focusInputTimestamp
-    submittedCommandFocusRecoveryGeneration = nil
-  }
-
-  func commitWorkspaceCommandFocus(
-    result: NativeFocusResult,
-    request: PendingWorkspaceFocus
-  ) {
-    guard pendingWorkspaceFocus?.commandGeneration == request.commandGeneration
-    else { return }
-    submittedWorkspaceFocusRequestID = nil
-    submittedWorkspaceFocusRequestTimestamp = nil
-    submittedWorkspaceFocusRecoveryGeneration = nil
-    if result == .frameSuperseded {
-      submittedWorkspaceFocusGeneration = nil
-      return
-    }
-    if result == .failed || result == .failedAfterMutation,
-      let nextRetryCount = nextCommandFocusRetryCount(
-        currentRetryCount: request.retryCount,
-        maximumRetryCount: 1,
-        requestGeneration: request.commandGeneration,
-        currentGeneration: commandGeneration,
-        requestedWindowID: request.requestedWindowID,
-        selectedWindowID: state.selectedWindowID(on: request.monitorID)
-      )
-    {
-      submittedWorkspaceFocusGeneration = nil
-      pendingWorkspaceFocus = PendingWorkspaceFocus(
-        monitorID: request.monitorID,
-        requestedWorkspaceID: request.requestedWorkspaceID,
-        previousWorkspaceID: request.previousWorkspaceID,
-        requestedWindowID: request.requestedWindowID,
-        restoresPreviousWorkspaceOnCancellation:
-          request.restoresPreviousWorkspaceOnCancellation,
-        commandGeneration: request.commandGeneration,
-        focusInputTimestamp: request.focusInputTimestamp,
-        cursorWarpInputTimestamp: request.cursorWarpInputTimestamp,
-        retryCount: nextRetryCount
-      )
-      return
-    }
-    pendingWorkspaceFocus = nil
-    submittedWorkspaceFocusGeneration = nil
-    guard !cancellationKeepsRequestedWindow(
-      request.requestedWindowID,
-      requestInputTimestamp: request.focusInputTimestamp
-    ) else { return }
-    guard let monitor = state.monitors.first(where: { $0.id == request.monitorID }),
-      let fallbackWorkspaceID = workspaceFocusCancellationFallback(
-        cancelledBeforeMutation:
-          result == .cancelled || result == .failed,
-        rollbackAfterMutation:
-          focusCompletionRequiresLogicalRollback(result),
-        requestGeneration: request.commandGeneration,
-        currentGeneration: self.commandGeneration,
-        requestedWorkspaceID: request.requestedWorkspaceID,
-        activeWorkspaceID: monitor.activeWorkspace,
-        previousWorkspaceID: request.previousWorkspaceID,
-        requestedWindowID: request.requestedWindowID,
-        selectedWindowID: state.selectedWindowID(on: request.monitorID),
-        restoresPreviousWorkspace:
-          request.restoresPreviousWorkspaceOnCancellation
-      )
-    else {
-      return
-    }
-    do {
-      try reduce(
-        .switchWorkspace(fallbackWorkspaceID),
-        on: request.monitorID,
-        state: &state
-      )
-    } catch {
-      return
-    }
-
-    activeMonitorID = request.monitorID
-    persistPlacements()
-    updateMenuBar()
-    synchronizeScrollOffsets(state: &state, viewports: viewportsByMonitor)
-    snapScrollOffsetsToTargets()
-    applyCurrentLayout(
-      asynchronousPositions: true,
-      updateVisibility: true,
-      positionTimeoutSeconds: 0.05,
-      stagesVisibleBeforeParking: true,
-      source: "workspace-focus-cancel"
-    )
-    needsDesktopSync = true
-  }
 }
