+++ Sources/DefiMacOS/AXFocusWriter.swift
@@ -6,236 +6,37 @@ import DefiCore
 import DefiModel
 import OSLog
 
-struct FocusInputGuard: @unchecked Sendable {
-  let tracker: UserInputTracker
-  let maximumTimestamp: TimeInterval
-}
-
-struct AsyncFocusRequest: @unchecked Sendable {
-  let element: AXUIElement
-  let application: AXUIElement
-  let processID: pid_t
-  let selectsSpecificWindow: Bool
-  let validatesSpecificWindowFocus: Bool
-  let activatesApplication: Bool
-  let inputGuard: FocusInputGuard?
-  let recoveryRequest: NativeFocusRecoveryRequest?
-}
-
-public enum NativeFocusResult: Equatable, Sendable {
-  case completed
-  case completedWithoutMutation
-  case frameSuperseded
-  case superseded
-  case supersededAfterMutation
-  case cancelled
-  case cancelledAfterMutation
-  case cancelledAfterInputMutation
-  case failed
-  case failedAfterMutation
-}
-
-public struct NativeFocusRequestID: Equatable, Sendable {
-  fileprivate let rawValue: UInt64
-}
-
-func focusRequestCanBeCancelled(
-  requestGeneration: UInt64,
-  latestGeneration: UInt64,
-  pendingGeneration: UInt64?,
-  activeGeneration: UInt64?
-) -> Bool {
-  requestGeneration == latestGeneration
-    && (pendingGeneration == requestGeneration
-      || activeGeneration == requestGeneration)
-}
-
-struct NativeFocusRecoveryFallback: Equatable, Sendable {
-  let windowID: WindowID?
-  let processID: pid_t?
-}
-
-struct NativeFocusRecoveryRequest: Equatable, Sendable {
-  let timestamp: TimeInterval
-  let excludingWindowID: WindowID
-  let excludingProcessID: pid_t
-  let fallback: NativeFocusRecoveryFallback?
-  let fallbackOnlyIfNoNewerInput: Bool
-
-  init(
-    timestamp: TimeInterval,
-    excludingWindowID: WindowID,
-    excludingProcessID: pid_t,
-    fallback: NativeFocusRecoveryFallback?,
-    fallbackOnlyIfNoNewerInput: Bool = false
-  ) {
-    self.timestamp = timestamp
-    self.excludingWindowID = excludingWindowID
-    self.excludingProcessID = excludingProcessID
-    self.fallback = fallback
-    self.fallbackOnlyIfNoNewerInput = fallbackOnlyIfNoNewerInput
-  }
-}
-
-func nativeFocusRecoveryFallbackTarget(
-  _ request: NativeFocusRecoveryRequest,
-  latestEventTimestamp: TimeInterval
-) -> UserInputTracker.FocusRecoveryTarget? {
-  guard request.fallbackOnlyIfNoNewerInput,
-    let fallback = request.fallback,
-    latestEventTimestamp <= request.timestamp
-  else {
-    return nil
-  }
-  return UserInputTracker.FocusRecoveryTarget(
-    timestamp: request.timestamp,
-    windowID: fallback.windowID,
-    processID: fallback.processID
-  )
-}
-
-struct NativeFocusCompletion: Equatable, Sendable {
-  let result: NativeFocusResult
-  let recoveryRequest: NativeFocusRecoveryRequest?
-}
-
-struct NativeFocusRecoveryTransfer: Equatable, Sendable {
-  let carried: NativeFocusRecoveryRequest?
-  let recovery: NativeFocusRecoveryRequest?
-}
-
-func transferredNativeFocusRecovery(
-  carried: NativeFocusRecoveryRequest?,
-  request: NativeFocusRecoveryRequest?,
-  result: NativeFocusResult,
-  generationCurrent: Bool
-) -> NativeFocusRecoveryTransfer {
-  switch result {
-  case .supersededAfterMutation, .cancelledAfterMutation:
-    return NativeFocusRecoveryTransfer(
-      carried: carried ?? request,
-      recovery: nil
-    )
-  case .cancelledAfterInputMutation:
-    guard generationCurrent else {
-      return NativeFocusRecoveryTransfer(carried: carried, recovery: nil)
-    }
-    return NativeFocusRecoveryTransfer(
-      carried: nil,
-      recovery: carried ?? request
-    )
-  case .frameSuperseded, .superseded, .cancelled, .failed,
-    .failedAfterMutation:
-    guard generationCurrent else {
-      return NativeFocusRecoveryTransfer(carried: carried, recovery: nil)
-    }
-    return NativeFocusRecoveryTransfer(
-      carried: nil,
-      recovery: result == .failedAfterMutation
-        ? carried ?? request
-        : carried
-    )
-  case .completed, .completedWithoutMutation:
-    guard generationCurrent else {
-      return NativeFocusRecoveryTransfer(carried: carried, recovery: nil)
-    }
-    return NativeFocusRecoveryTransfer(carried: nil, recovery: nil)
-  }
-}
-
-func resolvedNativeFocusResult(
-  mutationApplied: Bool,
-  generationCurrent: Bool,
-  inputCurrent: Bool,
-  cancelled: Bool,
-  focusSucceeded: Bool
-) -> NativeFocusResult {
-  if mutationApplied && generationCurrent && !inputCurrent {
-    return .cancelledAfterInputMutation
-  }
-  if mutationApplied && !generationCurrent {
-    return .supersededAfterMutation
-  }
-  if mutationApplied && (cancelled || !generationCurrent || !inputCurrent) {
-    return .cancelledAfterMutation
-  }
-  if !generationCurrent {
-    return .superseded
-  }
-  if !cancelled && generationCurrent && inputCurrent {
-    if focusSucceeded {
-      return mutationApplied ? .completed : .completedWithoutMutation
-    }
-    return mutationApplied ? .failedAfterMutation : .failed
-  }
-  return .cancelled
-}
-
-func nativeFocusRecoveryRequestForCompletion(
-  _ request: NativeFocusRecoveryRequest?,
-  result: NativeFocusResult,
-  explicitFallback: NativeFocusRecoveryFallback?
-) -> NativeFocusRecoveryRequest? {
-  guard let request else { return nil }
-  let fallback = explicitFallback ?? request.fallback
-  return NativeFocusRecoveryRequest(
-    timestamp: request.timestamp,
-    excludingWindowID: request.excludingWindowID,
-    excludingProcessID: request.excludingProcessID,
-    fallback: fallback,
-    fallbackOnlyIfNoNewerInput:
-      request.fallbackOnlyIfNoNewerInput
-      || explicitFallback != nil
-      || result == .failedAfterMutation
-  )
-}
-
-func focusMutationStateAfterActivation(
-  priorMutationApplied: Bool,
-  activationSucceeded: Bool
-) -> Bool {
-  priorMutationApplied || activationSucceeded
-}
-
-func specificWindowFocusWriteIsRequired(
-  requested: Bool,
-  validatesCurrentFocus: Bool,
-  targetIsFocused: Bool
-) -> Bool {
-  (requested || validatesCurrentFocus) && !targetIsFocused
-}
-
-private struct QueuedFocusRequest: @unchecked Sendable {
+struct QueuedFocusRequest: @unchecked Sendable {
   let generation: UInt64
   let request: AsyncFocusRequest
   let completion: @Sendable (NativeFocusCompletion) -> Void
 }
 
 final class AXFocusWriter: @unchecked Sendable {
-  private let queue = DispatchQueue(
+  let queue = DispatchQueue(
     label: "com.quentin.defi.ax-focus",
     qos: .userInitiated
   )
-  private let lock = NSLock()
-  private var pending: QueuedFocusRequest?
-  private var explicitlyCancelledGenerations = Set<UInt64>()
-  private var explicitCancellationFallbacks:
+  let lock = NSLock()
+  var pending: QueuedFocusRequest?
+  var explicitlyCancelledGenerations = Set<UInt64>()
+  var explicitCancellationFallbacks:
     [UInt64: NativeFocusRecoveryFallback] = [:]
-  private var activeProcessID: pid_t?
-  private var activeGeneration: UInt64?
-  private var needsRecoveryActivation = false
-  private var carriedFocusRecovery: NativeFocusRecoveryRequest?
-  private var latestGeneration: UInt64 = 0
-  private var minimumRecoveryGeneration: UInt64 = 0
-  private var appliedRecoveryResetGeneration: UInt64 = 0
-  private var running = false
-  private var lastDurationMS = 0.0
-  private var fastPathCount = 0
-  private var cancelledCount = 0
-  private var retryCount = 0
-  private var lastMainDurationMS = 0.0
-  private var lastRaiseDurationMS = 0.0
-  private var lastActivationDurationMS = 0.0
+  var activeProcessID: pid_t?
+  var activeGeneration: UInt64?
+  var needsRecoveryActivation = false
+  var carriedFocusRecovery: NativeFocusRecoveryRequest?
+  var latestGeneration: UInt64 = 0
+  var minimumRecoveryGeneration: UInt64 = 0
+  var appliedRecoveryResetGeneration: UInt64 = 0
+  var running = false
+  var lastDurationMS = 0.0
+  var fastPathCount = 0
+  var cancelledCount = 0
+  var retryCount = 0
+  var lastMainDurationMS = 0.0
+  var lastRaiseDurationMS = 0.0
+  var lastActivationDurationMS = 0.0
 
   @discardableResult
   func submit(
@@ -361,344 +162,4 @@ final class AXFocusWriter: @unchecked Sendable {
     )
   }
 
-  private func drain() {
-    while true {
-      lock.lock()
-      guard let queued = pending else {
-        running = false
-        lock.unlock()
-        return
-      }
-      pending = nil
-      activeProcessID = queued.request.processID
-      activeGeneration = queued.generation
-      lock.unlock()
-
-      let startedAt = ProcessInfo.processInfo.systemUptime
-      let request = queued.request
-      guard isCurrent(queued) else {
-        let generationCurrent = isCurrent(queued.generation)
-        lock.lock()
-        activeProcessID = nil
-        activeGeneration = nil
-        cancelledCount += 1
-        lock.unlock()
-        complete(
-          queued,
-          result: generationCurrent ? .cancelled : .superseded,
-          generationCurrent: generationCurrent
-        )
-        continue
-      }
-      var usedFastPath = false
-      var cancelled = false
-      var retried = false
-      var mainDurationMS = 0.0
-      var raiseDurationMS = 0.0
-      var activationDurationMS = 0.0
-      var focusMutationApplied = false
-      var windowSelectionSucceeded = false
-      var selectsSpecificWindow = request.selectsSpecificWindow
-      if selectsSpecificWindow || request.validatesSpecificWindowFocus {
-        selectsSpecificWindow = specificWindowFocusWriteIsRequired(
-          requested: selectsSpecificWindow,
-          validatesCurrentFocus: request.validatesSpecificWindowFocus,
-          targetIsFocused: isTargetFocused(
-            request.element,
-            application: request.application
-          )
-        )
-      }
-      if selectsSpecificWindow {
-        AXUIElementSetMessagingTimeout(request.application, 0.016)
-        AXUIElementSetMessagingTimeout(request.element, 0.016)
-        let mainStartedAt = ProcessInfo.processInfo.systemUptime
-        var mainResult = AXUIElementSetAttributeValue(
-          request.element,
-          kAXMainAttribute as CFString,
-          kCFBooleanTrue
-        )
-        focusMutationApplied = mainResult == .success
-        windowSelectionSucceeded = mainResult == .success
-        mainDurationMS =
-          (ProcessInfo.processInfo.systemUptime - mainStartedAt) * 1_000
-        cancelled = !isCurrent(queued)
-        var raiseResult = AXError.cannotComplete
-        if !cancelled, mainResult != .success {
-          let raiseStartedAt = ProcessInfo.processInfo.systemUptime
-          raiseResult = AXUIElementPerformAction(
-            request.element,
-            kAXRaiseAction as CFString
-          )
-          focusMutationApplied =
-            focusMutationApplied
-            || raiseResult == .success
-          windowSelectionSucceeded =
-            windowSelectionSucceeded
-            || raiseResult == .success
-          raiseDurationMS =
-            (ProcessInfo.processInfo.systemUptime - raiseStartedAt) * 1_000
-        }
-        cancelled = cancelled || !isCurrent(queued)
-        if !cancelled,
-          mainResult != .success && raiseResult != .success
-        {
-          retried = true
-          AXUIElementSetMessagingTimeout(request.application, 0.05)
-          AXUIElementSetMessagingTimeout(request.element, 0.05)
-          let retryMainStartedAt = ProcessInfo.processInfo.systemUptime
-          mainResult = AXUIElementSetAttributeValue(
-            request.element,
-            kAXMainAttribute as CFString,
-            kCFBooleanTrue
-          )
-          focusMutationApplied =
-            focusMutationApplied
-            || mainResult == .success
-          windowSelectionSucceeded =
-            windowSelectionSucceeded
-            || mainResult == .success
-          mainDurationMS +=
-            (ProcessInfo.processInfo.systemUptime - retryMainStartedAt) * 1_000
-          if isCurrent(queued), mainResult != .success {
-            let retryRaiseStartedAt = ProcessInfo.processInfo.systemUptime
-            raiseResult = AXUIElementPerformAction(
-              request.element,
-              kAXRaiseAction as CFString
-            )
-            focusMutationApplied =
-              focusMutationApplied
-              || raiseResult == .success
-            windowSelectionSucceeded =
-              windowSelectionSucceeded
-              || raiseResult == .success
-            raiseDurationMS +=
-              (ProcessInfo.processInfo.systemUptime - retryRaiseStartedAt) * 1_000
-          }
-          cancelled = !isCurrent(queued)
-        }
-        resetTimeouts(request)
-      } else {
-        usedFastPath = true
-        windowSelectionSucceeded = true
-      }
-      var activationAttempted = false
-      var activationSucceeded = true
-      let activationRequired =
-        cancelled || !isCurrent(queued)
-        ? nil
-        : activationRequirement(
-          requested: request.activatesApplication,
-          generation: queued.generation
-        )
-      if activationRequired == true {
-        activationAttempted = true
-        activationSucceeded = false
-        let activationStartedAt = ProcessInfo.processInfo.systemUptime
-        let system = AXUIElementCreateSystemWide()
-        let activationResult = AXUIElementSetAttributeValue(
-          system,
-          kAXFocusedApplicationAttribute as CFString,
-          request.application
-        )
-        if activationResult == .success {
-          activationSucceeded = true
-        } else if isCurrent(queued) {
-          activationSucceeded =
-            NSRunningApplication(processIdentifier: request.processID)?
-            .activate() == true
-        }
-        focusMutationApplied = focusMutationStateAfterActivation(
-          priorMutationApplied: focusMutationApplied,
-          activationSucceeded: activationSucceeded
-        )
-        activationDurationMS =
-          (ProcessInfo.processInfo.systemUptime - activationStartedAt) * 1_000
-      } else if activationRequired == nil {
-        cancelled = true
-      }
-      if activationAttempted, !isCurrent(queued) {
-        markRecoveryActivationNeeded()
-        cancelled = true
-      }
-      let durationMS =
-        (ProcessInfo.processInfo.systemUptime - startedAt) * 1_000
-      lock.lock()
-      lastDurationMS = durationMS
-      if usedFastPath { fastPathCount += 1 }
-      if cancelled { cancelledCount += 1 }
-      if retried { retryCount += 1 }
-      lastMainDurationMS = mainDurationMS
-      lastRaiseDurationMS = raiseDurationMS
-      lastActivationDurationMS = activationDurationMS
-      activeProcessID = nil
-      activeGeneration = nil
-      lock.unlock()
-      let generationCurrent = isCurrent(queued.generation)
-      let inputCurrent = inputGuardIsCurrent(request)
-        && !isExplicitlyCancelled(queued.generation)
-      let focusSucceeded =
-        windowSelectionSucceeded && activationSucceeded
-      let result = resolvedNativeFocusResult(
-        mutationApplied: focusMutationApplied,
-        generationCurrent: generationCurrent,
-        inputCurrent: inputCurrent,
-        cancelled: cancelled,
-        focusSucceeded: focusSucceeded
-      )
-      complete(
-        queued,
-        result: result,
-        generationCurrent: generationCurrent
-      )
-    }
-  }
-
-  private func complete(
-    _ queued: QueuedFocusRequest,
-    result: NativeFocusResult,
-    generationCurrent: Bool
-  ) {
-    lock.lock()
-    explicitlyCancelledGenerations.remove(queued.generation)
-    let explicitCancellationFallback = explicitCancellationFallbacks.removeValue(
-      forKey: queued.generation
-    )
-    let recoveryResetGeneration = minimumRecoveryGeneration
-    lock.unlock()
-    if appliedRecoveryResetGeneration < recoveryResetGeneration {
-      carriedFocusRecovery = nil
-      appliedRecoveryResetGeneration = recoveryResetGeneration
-    }
-    let discardsRecovery = queued.generation < recoveryResetGeneration
-    if discardsRecovery {
-      carriedFocusRecovery = nil
-      queued.completion(
-        NativeFocusCompletion(result: result, recoveryRequest: nil)
-      )
-      return
-    }
-    let recoveryRequest = nativeFocusRecoveryRequestForCompletion(
-      queued.request.recoveryRequest,
-      result: result,
-      explicitFallback: explicitCancellationFallback
-    )
-    let recoveryTransfer = transferredNativeFocusRecovery(
-      carried: carriedFocusRecovery,
-      request: recoveryRequest,
-      result: result,
-      generationCurrent: generationCurrent
-    )
-    carriedFocusRecovery = recoveryTransfer.carried
-    queued.completion(
-      NativeFocusCompletion(
-        result: result,
-        recoveryRequest: recoveryTransfer.recovery
-      )
-    )
-  }
-
-  private func isCurrent(_ generation: UInt64) -> Bool {
-    lock.lock()
-    defer { lock.unlock() }
-    return latestGeneration == generation
-  }
-
-  private func isCurrent(_ queued: QueuedFocusRequest) -> Bool {
-    lock.lock()
-    let explicitlyCancelled = explicitlyCancelledGenerations.contains(
-      queued.generation
-    )
-    lock.unlock()
-    guard !explicitlyCancelled, isCurrent(queued.generation) else {
-      return false
-    }
-    return inputGuardIsCurrent(queued.request)
-  }
-
-  private func isExplicitlyCancelled(_ generation: UInt64) -> Bool {
-    lock.lock()
-    defer { lock.unlock() }
-    return explicitlyCancelledGenerations.contains(generation)
-  }
-
-  private func inputGuardIsCurrent(_ request: AsyncFocusRequest) -> Bool {
-    guard let inputGuard = request.inputGuard else { return true }
-    return guardedFocusIsCurrent(
-      latestInputTimestamp: inputGuard.tracker.latestEventTimestamp,
-      maximumInputTimestamp: inputGuard.maximumTimestamp
-    )
-  }
-
-  private func activationRequirement(
-    requested: Bool,
-    generation: UInt64
-  ) -> Bool? {
-    lock.lock()
-    defer { lock.unlock() }
-    guard latestGeneration == generation else { return nil }
-    let required = requested || needsRecoveryActivation
-    needsRecoveryActivation = false
-    return required
-  }
-
-  private func markRecoveryActivationNeeded() {
-    lock.lock()
-    needsRecoveryActivation = true
-    lock.unlock()
-  }
-
-  private func resetTimeouts(_ request: AsyncFocusRequest) {
-    AXUIElementSetMessagingTimeout(request.element, 0)
-    AXUIElementSetMessagingTimeout(request.application, 0)
-  }
-
-  private func isTargetFocused(
-    _ element: AXUIElement,
-    application: AXUIElement
-  ) -> Bool {
-    AXUIElementSetMessagingTimeout(application, 0.016)
-    AXUIElementSetMessagingTimeout(element, 0.016)
-    defer {
-      AXUIElementSetMessagingTimeout(element, 0)
-      AXUIElementSetMessagingTimeout(application, 0)
-    }
-    // AXMain is structural app state, not proof of keyboard focus. Apps such as
-    // Kaku can leave a non-focused window main during rapid intra-app navigation.
-    return targetWindowFocusIsConfirmed(
-      readBoolean(element, attribute: kAXFocusedAttribute)
-    ) {
-      var focusedWindow: CFTypeRef?
-      guard AXUIElementCopyAttributeValue(
-        application,
-        kAXFocusedWindowAttribute as CFString,
-        &focusedWindow
-      ) == .success,
-        let focusedWindow,
-        CFGetTypeID(focusedWindow) == AXUIElementGetTypeID()
-      else {
-        return false
-      }
-      return CFEqual(focusedWindow, element)
-    }
-  }
-
-  private func readBoolean(
-    _ element: AXUIElement,
-    attribute: String
-  ) -> Bool? {
-    var rawValue: CFTypeRef?
-    guard
-      AXUIElementCopyAttributeValue(
-        element,
-        attribute as CFString,
-        &rawValue
-      ) == .success,
-      let rawValue,
-      CFGetTypeID(rawValue) == CFBooleanGetTypeID()
-    else {
-      return nil
-    }
-    return CFBooleanGetValue((rawValue as! CFBoolean))
-  }
 }
